+++ RandomSkunk.Results/Operations/GetValueOr.cs
@@ -1,6 +1,6 @@
 namespace RandomSkunk.Results;
 
-/// <content> Defines the <c>GetValueOr</c> methods. </content>
+/// <content> Defines the <c>GetValueOr</c> and <c>GetValueOrDefault</c> methods. </content>
 public partial struct Result<T>
 {
     /// <summary>
@@ -28,9 +28,17 @@ public partial struct Result<T>
 
         return _outcome == _successOutcome ? _value! : getFallbackValue();
     }
+
+    /// <summary>
+    /// Gets the value of the <c>Success</c> result, or the default value of type <typeparamref name="T"/> if it is a <c>Fail</c>
+    /// result.
+    /// </summary>
+    /// <returns>The value of this result if this is a <c>Success</c> result; otherwise, the default value of type
+    ///     <typeparamref name="T"/>.</returns>
+    public T? GetValueOrDefault() => GetValueOr((T?)default);
 }
 
-/// <content> Defines the <c>GetValueOr</c> methods. </content>
+/// <content> Defines the <c>GetValueOr</c> and <c>GetValueOrDefault</c> methods. </content>
 public partial struct Maybe<T>
 {
     /// <inheritdoc cref="Result{T}.GetValueOr(T)"/>
@@ -46,9 +54,12 @@ public partial struct Maybe<T>
 
         return _outcome == _successOutcome ? _value! : getFallbackValue();
     }
+
+    /// <inheritdoc cref="Result{T}.GetValueOrDefault"/>
+    public T? GetValueOrDefault() => GetValueOr((T?)default);
 }
 
-/// <content> Defines the <c>GetValueOr</c> extension methods. </content>
+/// <content> Defines the <c>GetValueOr</c> and <c>GetValueOrDefault</c> extension methods. </content>
 public static partial class ResultExtensions
 {
     #pragma warning disable CS1573 // Parameter has no matching param tag in the XML comment (but other parameters do)
@@ -65,6 +76,12 @@ public static partial class ResultExtensions
     public static async Task<T?> GetValueOr<T>(this Task<Result<T>> sourceResult, Func<T?> getFallbackValue) =>
         (await sourceResult.ConfigureAwait(false)).GetValueOr(getFallbackValue);
 
+    /// <inheritdoc cref="Result{T}.GetValueOrDefault"/>
+    /// <typeparam name="T">The type of the source result value.</typeparam>
+    /// <param name="sourceResult">The source result.</param>
+    public static async Task<T?> GetValueOrDefault<T>(this Task<Result<T>> sourceResult) =>
+        (await sourceResult.ConfigureAwait(false)).GetValueOrDefault();
+
     /// <inheritdoc cref="GetValueOr{T}(Task{Result{T}}, T)"/>
     public static async Task<T?> GetValueOr<T>(this Task<Maybe<T>> sourceResult, T? fallbackValue) =>
         (await sourceResult.ConfigureAwait(false)).GetValueOr(fallbackValue);
@@ -73,5 +90,9 @@ public static partial class ResultExtensions
     public static async Task<T?> GetValueOr<T>(this Task<Maybe<T>> sourceResult, Func<T?> getFallbackValue) =>
         (await sourceResult.ConfigureAwait(false)).GetValueOr(getFallbackValue);
 
-    #pragma warning restore CS1573 // Parameter has no matching param tag in the XML comment (but other parameters do)
+    /// <inheritdoc cref="GetValueOrDefault{T}(Task{Result{T}})"/>
+    public static async Task<T?> GetValueOrDefault<T>(this Task<Maybe<T>> sourceResult) =>
+        (await sourceResult.ConfigureAwait(false)).GetValueOrDefault();
+
+#pragma warning restore CS1573 // Parameter has no matching param tag in the XML comment (but other parameters do)
 }